auth-proxy/pyproject.toml
[project]
name = "obsidian-auth-proxy"
version = "1.0.0"
description = "Custom JWT authentication proxy for Obsidian LiveSync with per-device token management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "PyJWT>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
